requirements.txt
Pillow==9.2.0
Pillow==9.5.0
Requests==2.30.0
streamlit==1.14.0
streamlit_lottie==0.0.3
